+++ src/Testing/RuntimeOutcomeInvalid.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Doria\Baton\Testing;
+
+final class RuntimeOutcomeInvalid extends \RuntimeException
+{
+}